# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PromptHub API Documentation",
        "description": "API completa para gerenciamento de prompts de IA e agentes AI. Sistema que permite autenticacao via email/senha ou Google OAuth, cadastro e gerenciamento de agentes de IA personalizados, envio de prompts para OpenAI e Google Gemini.",
        "contact": {
            "name": "PromptHub Support",
            "email": "suporte@prompthub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://dflix-prompthub-backend.a8zera.easypanel.host/api",
            "description": "Production Server (Easypanel)"
        },
        {
            "url": "https://api-prompthub.marcioleite.cloud/api",
            "description": "Production Server (Alternative)"
        }
    ],
    "paths": {
        "/agents": {
            "get": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Listar agentes de IA",
                "description": "Retorna todos os agentes de IA do usuario autenticado ordenados por data de criacao (mais recentes primeiro).",
                "operationId": "listAgents",
                "responses": {
                    "200": {
                        "description": "Lista de agentes retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Assistente de Codigo Python"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Agente especializado em revisar e otimizar codigo Python"
                                            },
                                            "ai_provider": {
                                                "type": "string",
                                                "example": "openai"
                                            },
                                            "ai_model": {
                                                "type": "string",
                                                "example": "gpt-4",
                                                "nullable": true
                                            },
                                            "system_prompt": {
                                                "type": "string",
                                                "example": "Voce e um especialista em Python...",
                                                "nullable": true
                                            },
                                            "instructions": {
                                                "type": "string",
                                                "example": "Sempre siga as boas praticas PEP 8...",
                                                "nullable": true
                                            },
                                            "configuration": {
                                                "type": "object",
                                                "example": {
                                                    "temperature": 0.5,
                                                    "max_tokens": 2000
                                                },
                                                "nullable": true
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Criar novo agente de IA",
                "description": "Cria um agente de IA personalizado com instrucoes e configuracoes especificas para o usuario autenticado.",
                "operationId": "createAgent",
                "requestBody": {
                    "description": "Dados do agente",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "ai_provider"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do agente",
                                        "type": "string",
                                        "example": "Assistente de Codigo Python"
                                    },
                                    "description": {
                                        "description": "Descricao do agente",
                                        "type": "string",
                                        "example": "Agente especializado em revisar codigo Python",
                                        "nullable": true
                                    },
                                    "ai_provider": {
                                        "description": "Provedor de IA",
                                        "type": "string",
                                        "enum": [
                                            "openai",
                                            "gemini"
                                        ],
                                        "example": "openai"
                                    },
                                    "ai_model": {
                                        "description": "Modelo especifico da IA",
                                        "type": "string",
                                        "example": "gpt-4",
                                        "nullable": true
                                    },
                                    "system_prompt": {
                                        "description": "Prompt de sistema para contexto",
                                        "type": "string",
                                        "example": "Voce e um especialista em Python...",
                                        "nullable": true
                                    },
                                    "instructions": {
                                        "description": "Instrucoes adicionais",
                                        "type": "string",
                                        "example": "Sempre siga PEP 8...",
                                        "nullable": true
                                    },
                                    "configuration": {
                                        "description": "Configuracoes personalizadas",
                                        "type": "object",
                                        "example": {
                                            "temperature": 0.5,
                                            "max_tokens": 2000
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agente criado com sucesso!"
                                        },
                                        "agent": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Assistente de Codigo Python"
                                                },
                                                "ai_provider": {
                                                    "type": "string",
                                                    "example": "openai"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/agents/{id}": {
            "get": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Obter agente especifico",
                "description": "Retorna os detalhes completos de um agente de IA especifico do usuario autenticado.",
                "operationId": "getAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agente retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Assistente de Codigo Python"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "ai_provider": {
                                            "type": "string",
                                            "example": "openai"
                                        },
                                        "system_prompt": {
                                            "type": "string"
                                        },
                                        "configuration": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Agente nao encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\AiAgent]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Atualizar agente",
                "description": "Atualiza as informacoes de um agente de IA existente do usuario autenticado.",
                "operationId": "updateAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos a serem atualizados",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Assistente de Codigo Python Avancado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "ai_provider": {
                                        "type": "string",
                                        "enum": [
                                            "openai",
                                            "gemini"
                                        ]
                                    },
                                    "ai_model": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "system_prompt": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "instructions": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "configuration": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agente atualizado com sucesso!"
                                        },
                                        "agent": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Agente nao encontrado"
                    },
                    "422": {
                        "description": "Erro de validacao"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Deletar agente",
                "description": "Remove permanentemente um agente de IA do usuario autenticado.",
                "operationId": "deleteAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agente deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agente deletado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Agente nao encontrado"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/agents/{id}/prompt": {
            "post": {
                "tags": [
                    "AI Agents"
                ],
                "summary": "Enviar prompt para agente especifico",
                "description": "Envia um prompt para um agente de IA especifico. O agente aplica seu system_prompt, instrucoes e configuracoes ao processar o prompt.",
                "operationId": "sendPromptToAgent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do agente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do prompt",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token",
                                    "prompt"
                                ],
                                "properties": {
                                    "api_token": {
                                        "description": "Token de API do provedor configurado no agente",
                                        "type": "string",
                                        "example": "sk-proj-abc123..."
                                    },
                                    "prompt": {
                                        "description": "Texto do prompt do usuario",
                                        "type": "string",
                                        "minLength": 10,
                                        "example": "Revise este codigo Python e sugira melhorias"
                                    },
                                    "file_content": {
                                        "description": "Conteudo de arquivo ou dados adicionais",
                                        "type": "string",
                                        "example": "def soma(a, b):\\n    return a+b",
                                        "nullable": true
                                    },
                                    "additional_data": {
                                        "description": "Informacoes extras para contexto",
                                        "type": "string",
                                        "example": "O codigo deve seguir PEP 8",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta do agente recebida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "response": {
                                            "type": "string",
                                            "example": "Aqui estao as sugestoes de melhoria para seu codigo..."
                                        },
                                        "agent": {
                                            "type": "string",
                                            "example": "Assistente de Codigo Python"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Agente nao encontrado"
                    },
                    "422": {
                        "description": "Erro de validacao"
                    },
                    "500": {
                        "description": "Erro ao processar prompt"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login de usuario",
                "description": "Autentica usuario com email e senha. Retorna token JWT para autorizacao.",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciais do usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Joao Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@exemplo.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais invalidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}